requirements.txt
Flask==2.3.3
pytest==8.3.5
pytest-flask==1.3.0
pytest-cov==4.1.0
Werkzeug==2.3.7
PyJWT==2.8.0
coverage==7.3.2
